+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/STPAnalyticsClient+PaymentSheet.swift
@@ -366,8 +366,6 @@ extension PaymentSheet.Configuration {
         payload["save_payment_method_opt_in_behavior"] = savePaymentMethodOptInBehavior.description
         payload["appearance"] = appearance.analyticPayload
         payload["billing_details_collection_configuration"] = billingDetailsCollectionConfiguration.analyticPayload
-        // Only grab the first 10 as a safeguard, since `externalPaymentMethods` is configured by the merchant and theoretically ~unbounded
-        payload["external_payment_methods"] = externalPaymentMethodConfiguration?.externalPaymentMethods.prefix(10)
         return payload
     }
 }